Refactor MoveableEntity move mechanics; simplify/reduce coupling between ME subtypes by adding moveDirection field to ME and making movement direction the main cause of an ME's movement

+++ src/World/Entities/Collidables/Organisms/Humanoid/Player/PlayerInputManager.cpp
@@ -45,7 +45,7 @@ void PlayerInputManager::handleMovement() {
     const auto sPressed = sf::Keyboard::isKeyPressed(Key::S);
     const auto wPressed = sf::Keyboard::isKeyPressed(Key::W);
 
-    sf::Vector2i newMoveDir;
+    sf::Vector2f newMoveDir;
 
     if (wPressed) {
         newMoveDir.y = -1;
